requirements.txt
pyyaml>=6.0
jsonschema>=4.17
pytest>=7.0
